Move around util files
