--- powergrid/core/features/__init__.py
@@ -1,34 +1,0 @@
-from powergrid.core.features.electrical import ElectricalBasePh
-from powergrid.core.features.thermal import ThermalLoading
-from powergrid.core.features.tap_changer import TapChangerPh
-from powergrid.core.features.var import ShuntCapacitorBlock#, VoltVarCurve
-# from powergrid.core.features.protection import ProtectionRelay, Switchgear
-from powergrid.core.features.inverters import InverterBasedSource
-from powergrid.core.features.storage import StorageBlock
-# from powergrid.core.features.ev import EVBlock
-from powergrid.core.features.connection import PhaseConnection
-from powergrid.core.features.status import StatusBlock
-
-
-__all__ = [
-    # electrical
-    "ElectricalBasePh",
-    # asset
-    "ThermalLoading",
-    # tap_changer
-    "TapChangerPh",
-    # var
-    "ShuntCapacitorBlock",# "VoltVarCurve",
-    # # protection
-    # "ProtectionRelay", "Switchgear",
-    # inverters
-    "InverterBasedSource",
-    # storage
-    "StorageBlock",
-    # # ev
-    # "EVBlock",
-    # connection
-    "PhaseConnection",
-    # status
-    "StatusBlock",
-]
