+++ powergrid/core/features/thermal.py
@@ -2,9 +2,11 @@ from dataclasses import dataclass, field, asdict
 from typing import Optional, List, Dict
 import numpy as np
 
-from powergrid.core.utils.typing import Array, FeatureProvider
-from powergrid.core.utils.registry import provider
-from powergrid.core.utils.utils import _as_f32
+from powergrid.utils.typing import Array
+from powergrid.core.features.base import FeatureProvider
+
+from powergrid.utils.registry import provider
+from powergrid.utils.array_utils import _as_f32
 from powergrid.core.state import PhaseModel, PhaseSpec
