+++ powergrid/devices/compensation.py
@@ -1,4 +1,3 @@
-from builtins import float as builtin_float
 import numpy as np
 from typing import Any, Dict, Optional, List
 from dataclasses import dataclass
@@ -6,9 +5,9 @@ from dataclasses import dataclass
 from powergrid.core.policies import Policy
 from powergrid.core.typing import Array, FeatureProvider
 from powergrid.core.state import PhaseModel, PhaseSpec
-from powergrid.core.registry import provider
-from ..agents.device_agent import DeviceAgent
-from ..core.protocols import NoProtocol, Protocol
+from powergrid.utils.registry import provider
+from powergrid.agents.device_agent import DeviceAgent
+from powergrid.core.protocols import NoProtocol, Protocol
 
 
 # Create provider for step-based discrete state
@@ -71,9 +70,9 @@ class Shunt(DeviceAgent):
         self.type = "SCB"
         self.name = name
         self.bus = bus
-        self.q_mvar = builtin_float(q_mvar)
+        self.q_mvar = float(q_mvar)
         self.max_step = int(max_step)
-        self.switching_cost = builtin_float(switching_cost)
+        self.switching_cost = float(switching_cost)
         self._last_step = 0
         
         super().__init__(
@@ -87,7 +86,6 @@ class Shunt(DeviceAgent):
         # discrete steps: 0..max_step
         self.action.ncats = self.max_step + 1
         self.action.dim_d = 1
-        
         self.action.sample()
 
     def set_device_state(self):
@@ -107,7 +105,7 @@ class Shunt(DeviceAgent):
 
     def update_cost_safety(self) -> None:
         changed = int(getattr(self, "_current_step", 0) != getattr(self, "_last_step", 0))
-        self.cost = builtin_float(self.switching_cost * changed)
+        self.cost = float(self.switching_cost * changed)
         self.safety = 0.0
         self._last_step = getattr(self, "_current_step", self._last_step)
