+++ powergrid/utils/typing.py
@@ -1,0 +1,9 @@
+import numpy as np
+
+
+Array = np.ndarray
+FloatArray = np.ndarray
+IntArray = np.ndarray
+
+
+
