+++ tests/core/features/test_inverters.py
@@ -1,7 +1,7 @@
 import numpy as np
 import pytest
 
-from powergrid.core.utils.phase import PhaseModel, PhaseSpec
+from powergrid.utils.phase import PhaseModel, PhaseSpec
 from powergrid.core.features import InverterBasedSource
